Use block and state slot to get ForkName and type
