+++ packages/lodestar/src/db/repositories/blockArchiveIndex.ts
