+++ packages/lodestar/src/db/repositories/stateArchive.ts
@@ -1,22 +1,30 @@
-/* eslint-disable @typescript-eslint/no-unsafe-return */
-/* eslint-disable @typescript-eslint/no-unsafe-call */
-import {CompositeType, ContainerType, TreeBacked} from "@chainsafe/ssz";
+import {ContainerType, TreeBacked} from "@chainsafe/ssz";
 import {Epoch, Root, Slot, allForks} from "@chainsafe/lodestar-types";
 import {IBeaconConfig} from "@chainsafe/lodestar-config";
 import {bytesToInt} from "@chainsafe/lodestar-utils";
 import {IDatabaseController, Bucket, Repository} from "@chainsafe/lodestar-db";
-import {getRootIndexKey, storeRootIndex} from "./db-index";
+import {getStateTypeFromBytes, getStateTypeFromState} from "./utils/multifork";
+import {getRootIndexKey, storeRootIndex} from "./stateArchiveIndex";
+
+/* eslint-disable @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-call */
 
 export class GenericStateArchiveRepository extends Repository<Slot, TreeBacked<allForks.BeaconState>> {
-  constructor(
-    config: IBeaconConfig,
-    db: IDatabaseController<Buffer, Buffer>,
-    bucket: Bucket,
-    type: ContainerType<allForks.BeaconState>
-  ) {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    super(config, db, bucket, type as any);
+  constructor(config: IBeaconConfig, db: IDatabaseController<Buffer, Buffer>) {
+    const type = (config.types.phase0.BeaconState as unknown) as ContainerType<TreeBacked<allForks.BeaconState>>; // Pick some type but won't be used
+    super(config, db, Bucket.allForks_stateArchive, type);
   }
+
+  // Overrides for multi-fork
+
+  encodeValue(value: allForks.BeaconState): Buffer {
+    return getStateTypeFromState(this.config, value).serialize(value) as Buffer;
+  }
+
+  decodeValue(data: Buffer): TreeBacked<allForks.BeaconState> {
+    return getStateTypeFromBytes(this.config, data).createTreeBackedFromBytes(data);
+  }
+
+  // Handle key as slot
 
   async put(key: Slot, value: TreeBacked<allForks.BeaconState>): Promise<void> {
     await Promise.all([super.put(key, value), storeRootIndex(this.db, key, value.hashTreeRoot())]);
@@ -30,9 +38,7 @@ export class GenericStateArchiveRepository extends Repository<Slot, TreeBacked<a
     return bytesToInt((super.decodeKey(data) as unknown) as Uint8Array, "be");
   }
 
-  decodeValue(data: Buffer): TreeBacked<allForks.BeaconState> {
-    return ((this.type as unknown) as CompositeType<allForks.BeaconState>).createTreeBackedFromBytes(data);
-  }
+  // Index Root -> Slot
 
   async getByRoot(stateRoot: Root): Promise<TreeBacked<allForks.BeaconState> | null> {
     const slot = await this.getSlotByRoot(stateRoot);
@@ -44,9 +50,6 @@ export class GenericStateArchiveRepository extends Repository<Slot, TreeBacked<a
 
   private async getSlotByRoot(root: Root): Promise<Slot | null> {
     const value = await this.db.get(getRootIndexKey(root));
-    if (value) {
-      return bytesToInt(value, "be");
-    }
-    return null;
+    return value && bytesToInt(value, "be");
   }
 }
