+++ packages/lodestar/src/db/repositories/stateArchiveIndex.ts
